+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 Changed the io.aviso.macros namespace to be io.aviso.toolchest.macros.
 
-Fixes #1
+Added the io.aviso.toolchest.exceptions and collections namespaces.
 
 ## 0.1.0 - 17 Nov 2014