+++ CHANGES.md
@@ -1,4 +1,4 @@
-## 0.1.1 - UNRELEASED
+## 0.1.1 - 25 Nov 2014
 
 Changed the io.aviso.macros namespace to be io.aviso.toolchest.macros.